refactor(machoanalyzer): extract arm64util package and address review comments
- Create internal/arm64util with BackwardScanX16, IsControlFlowInstruction,
  WritesX16NotMovzMovk; both libccache and machoanalyzer now import from
  arm64util, eliminating the duplicated ARM64 instruction decoding logic
- Add upper-bound check (panic on out-of-bounds) before uint64→int cast in
  scanSVCWithX16, satisfying gosec G115 without a nolint comment
- Use fn.Entry + uint64(offset) instead of uint64(int64(fn.Entry) + offset)
  to avoid risky intermediate int64 conversion for large virtual addresses
- Add arm64util to depguard allowlist in .golangci.yml

+++ internal/runner/security/machoanalyzer/pclntab_macho.go
@@ -85,8 +85,8 @@ func ParseMachoPclntab(f *macho.File) (map[string]MachoPclntabFunc, error) {
 		for name, fn := range functions {
 			functions[name] = MachoPclntabFunc{
 				Name:  fn.Name,
-				Entry: uint64(int64(fn.Entry) + offset), //nolint:gosec // G115: offset is bounded by binary size, no overflow risk
-				End:   uint64(int64(fn.End) + offset),   //nolint:gosec // G115: offset is bounded by binary size, no overflow risk
+				Entry: fn.Entry + uint64(offset), //nolint:gosec // G115: offset verified positive in detectMachoPclntabOffset
+				End:   fn.End + uint64(offset),   //nolint:gosec // G115: offset verified positive in detectMachoPclntabOffset
 			}
 		}
 	}
